Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    float a;
-    float b;
+    int a;
+    int b;
     char ch;
-    scanf("%f %f %c",&a,&b,&ch);
+    scanf("%d %d %c",&a,&b,&ch);
     switch(ch){
         case '+':
             printf("%d",a+b);
